First pass at extracting RepositorySession
