+++ packages/server/src/fhir/repository/repository-connection.ts
@@ -1,0 +1,392 @@
+// SPDX-FileCopyrightText: Copyright Orangebot, Inc. and Medplum contributors
+// SPDX-License-Identifier: Apache-2.0
+import type { OperationOutcomeError } from '@medplum/core';
+import { normalizeErrorString, sleep } from '@medplum/core';
+import { RepositoryMode } from '@medplum/fhir-router';
+import type { Pool, PoolClient } from 'pg';
+import { getConfig } from '../../config/loader';
+import { DatabaseMode, getDatabasePool } from '../../database';
+import { getLogger } from '../../logger';
+import type { TransactionIsolationLevel } from '../sql';
+import { isRetryableTransactionError, normalizeDatabaseError } from '../sql';
+
+const defaultTransactionAttempts = 2;
+const defaultExpBackoffBaseDelayMs = 50;
+
+type CallbackFrame = {
+  pre: number;
+  post: number;
+};
+
+type RepositoryConnectionOptions = {
+  client?: PoolClient;
+  ownsClient?: boolean;
+  mode?: RepositoryMode;
+};
+
+/**
+ * Shared database-session state for one or more Repository facades.
+ *
+ * Any repositories that can share a PoolClient should share a RepositoryConnection
+ * so transaction depth, savepoints, callbacks, and cache deferral decisions
+ * cannot diverge from the underlying Postgres transaction.
+ */
+export class RepositoryConnection implements Disposable {
+  private conn?: PoolClient;
+  private readonly ownsClient: boolean;
+  private transactionDepth = 0;
+  private closed = false;
+  mode: RepositoryMode;
+
+  private preCommitCallbacks: (() => Promise<void>)[] = [];
+  private postCommitCallbacks: (() => Promise<void>)[] = [];
+  private callbackStack: CallbackFrame[] = [];
+
+  constructor(options?: RepositoryConnectionOptions) {
+    this.conn = options?.client;
+    this.ownsClient = options?.ownsClient ?? true;
+    this.mode = options?.mode ?? RepositoryMode.WRITER;
+  }
+
+  static fromClient(client: PoolClient, options?: { ownsClient?: boolean }): RepositoryConnection {
+    return new RepositoryConnection({
+      client,
+      ownsClient: options?.ownsClient ?? false,
+    });
+  }
+
+  isInTransaction(): boolean {
+    return this.transactionDepth > 0;
+  }
+
+  hasConnection(): boolean {
+    return !!this.conn;
+  }
+
+  /**
+   * Returns a database client.
+   * Use this method when you don't care if you're in a transaction or not.
+   * For example, use this method for "read by ID".
+   * The return value can either be a pool client or a pool.
+   * If in a transaction, then returns the transaction client (PoolClient).
+   * Otherwise, returns the pool (Pool).
+   * @param mode - The database mode.
+   * @returns The database client.
+   */
+  getDatabaseClient(mode: DatabaseMode): Pool | PoolClient {
+    this.assertNotClosed();
+    if (this.conn) {
+      // If in a transaction, then use the transaction client.
+      return this.conn;
+    }
+    if (mode === DatabaseMode.WRITER) {
+      // If we ever use a writer, then all subsequent operations must use a writer.
+      this.mode = RepositoryMode.WRITER;
+    }
+    return getDatabasePool(this.mode === RepositoryMode.WRITER ? DatabaseMode.WRITER : mode);
+  }
+
+  /**
+   * Returns a proper database connection.
+   * Unlike getDatabaseClient(), this method always returns a PoolClient.
+   * @param mode - The database mode.
+   * @returns Database connection.
+   */
+  private async getConnection(mode: DatabaseMode): Promise<PoolClient> {
+    this.assertNotClosed();
+    this.conn ??= await getDatabasePool(mode).connect();
+    return this.conn;
+  }
+
+  /**
+   * Releases the database connection.
+   * Include an error to remove the connection from the pool.
+   * See: https://github.com/brianc/node-postgres/blob/master/packages/pg-pool/index.js#L333
+   * @param err - Optional error to remove the connection from the pool.
+   */
+  private releaseConnection(err?: boolean | Error): void {
+    if (!this.conn) {
+      return;
+    }
+    if (this.ownsClient) {
+      try {
+        this.conn.release(err);
+      } catch (releaseErr) {
+        // pg-pool throws if release() is called twice (e.g. socket already errored out).
+        // We've done our part; just log and move on.
+        getLogger().warn('Error releasing database client', { err: normalizeErrorString(releaseErr) });
+      }
+      this.conn = undefined;
+    } else if (err) {
+      // Shared connection is known to be dead. Drop our reference so we don't reuse it.
+      // The owner of the connection is responsible for the actual release.
+      this.conn = undefined;
+    }
+  }
+
+  async withTransaction<TResult>(
+    callback: (client: PoolClient) => Promise<TResult>,
+    options?: { serializable: boolean }
+  ): Promise<TResult> {
+    const config = getConfig();
+    const transactionAttempts = config.transactionAttempts ?? defaultTransactionAttempts;
+    let error: OperationOutcomeError | undefined;
+    for (let attempt = 0; attempt < transactionAttempts; attempt++) {
+      const attemptStartTime = Date.now();
+      try {
+        const client = await this.beginTransaction(options?.serializable ? 'SERIALIZABLE' : undefined);
+        const result = await callback(client);
+        await this.commitTransaction();
+        if (attempt > 0) {
+          getLogger().info('Completed transaction', {
+            attempt,
+            attemptDurationMs: Date.now() - attemptStartTime,
+            transactionAttempts,
+            serializable: options?.serializable ?? false,
+          });
+        }
+        return result;
+      } catch (err) {
+        const operationOutcomeError = normalizeDatabaseError(err);
+        // Assigning here and throwing below is necessary to satisfy TypeScript
+        error = operationOutcomeError;
+
+        // Ensure transaction is rolled back before attempting any retry
+        await this.rollbackTransaction(operationOutcomeError);
+        if (this.transactionDepth || !isRetryableTransactionError(operationOutcomeError)) {
+          break; // Fall through to throw statement outside of the loop
+        }
+      } finally {
+        this.endTransaction();
+      }
+
+      const attemptDurationMs = Date.now() - attemptStartTime;
+
+      if (attempt + 1 < transactionAttempts) {
+        const baseDelayMs = config.transactionExpBackoffBaseDelayMs ?? defaultExpBackoffBaseDelayMs;
+        // Attempts are 0-indexed, so first wait after first attempt will be somewhere between 75% and 125% of baseDelayMs
+        // This calculation results in something like this for the default values:
+        // Between attempt 0 and 1: 50 * (2^0) = 50 * [0.75, 1.25] = **[37.5, 63.5] ms**
+        // Between attempt 1 and 2: 50 * (2^1) = 100 * [0.75, 1.25] = **[75, 125] ms**
+        // etc...
+        const delayMs = Math.ceil(baseDelayMs * 2 ** attempt * (0.75 + Math.random() * 0.5));
+        getLogger().info('Retrying transaction', {
+          attempt,
+          attemptDurationMs,
+          transactionAttempts,
+          serializable: options?.serializable ?? false,
+          delayMs,
+          baseDelayMs,
+        });
+        await sleep(delayMs);
+      } else {
+        getLogger().info('Transaction failed final attempt', {
+          attempt,
+          attemptDurationMs,
+          transactionAttempts,
+          serializable: options?.serializable ?? false,
+        });
+      }
+    }
+
+    // Cannot be undefined: either the function returns normally from the `try` block,
+    // or `error` is assigned at top of `catch` block before reaching this line
+    throw error;
+  }
+
+  private async beginTransaction(isolationLevel: TransactionIsolationLevel = 'REPEATABLE READ'): Promise<PoolClient> {
+    this.assertNotClosed();
+    this.transactionDepth++;
+    this.pushCallbackFrame();
+    const conn = await this.getConnection(DatabaseMode.WRITER);
+    if (this.transactionDepth === 1) {
+      await conn.query('BEGIN ISOLATION LEVEL ' + isolationLevel);
+    } else {
+      await conn.query('SAVEPOINT sp' + this.transactionDepth);
+    }
+    return conn;
+  }
+
+  private async commitTransaction(): Promise<void> {
+    this.assertInTransaction();
+    const conn = await this.getConnection(DatabaseMode.WRITER);
+    if (this.transactionDepth === 1) {
+      await this.processPreCommit();
+      await conn.query('COMMIT');
+      this.transactionDepth--;
+      this.releaseConnection();
+      this.clearCallbackStack();
+      await this.processPostCommit();
+    } else {
+      // If RELEASE SAVEPOINT fails (e.g. transaction in aborted state), let the error propagate.
+      // withTransaction's catch will invoke rollbackTransaction, which can run ROLLBACK TO SAVEPOINT
+      // even against an aborted transaction to recover — aborting here would discard work the outer
+      // scope can still commit. rollbackTransaction's own catch handles the truly-dead-connection case.
+      await conn.query('RELEASE SAVEPOINT sp' + this.transactionDepth);
+      this.transactionDepth--;
+      this.popCallbackFrame();
+    }
+  }
+
+  private async rollbackTransaction(error: Error): Promise<void> {
+    // Tolerate being called after state has already been reset (e.g. when a prior
+    // cleanup path in commit/rollback fully aborted the transaction on a dead connection).
+    if (this.transactionDepth <= 0) {
+      return;
+    }
+    const conn = await this.getConnection(DatabaseMode.WRITER);
+    const isOuter = this.transactionDepth === 1;
+    try {
+      if (isOuter) {
+        await conn.query('ROLLBACK');
+      } else {
+        await conn.query('ROLLBACK TO SAVEPOINT sp' + this.transactionDepth);
+      }
+    } catch (rollbackErr) {
+      // ROLLBACK itself failed — connection is almost certainly dead (e.g. killed by
+      // idle_in_transaction_session_timeout). Pass the original triggering error to
+      // abortTransaction so the client is released with the right root cause.
+      getLogger().warn('Error rolling back transaction', {
+        err: normalizeErrorString(rollbackErr),
+        originalErr: normalizeErrorString(error),
+      });
+      this.abortTransaction(error);
+      return;
+    }
+    this.transactionDepth--;
+    this.truncateCommitCallbacks();
+    if (isOuter) {
+      this.releaseConnection(error);
+    }
+  }
+
+  /**
+   * Fully abort the transaction hierarchy: reset depth, drop pending pre/post-commit
+   * callbacks, and release the connection with an error so pg-pool discards it.
+   * Invoked from commit/rollback error paths when further recovery on the current
+   * connection is not possible.
+   * @param err - The error that triggered the abort; forwarded to `release()`.
+   */
+  private abortTransaction(err: Error): void {
+    this.transactionDepth = 0;
+    this.clearCallbackStack();
+    this.releaseConnection(err);
+  }
+
+  private endTransaction(): void {
+    if (this.transactionDepth === 0) {
+      this.releaseConnection();
+    }
+  }
+
+  private assertInTransaction(): void {
+    if (this.transactionDepth <= 0) {
+      throw new Error('Not in transaction');
+    }
+  }
+
+  async ensureInTransaction<TResult>(callback: (client: PoolClient) => Promise<TResult>): Promise<TResult> {
+    if (this.transactionDepth) {
+      const client = await this.getConnection(DatabaseMode.WRITER);
+      return callback(client);
+    } else {
+      return this.withTransaction(callback);
+    }
+  }
+
+  async preCommit(fn: () => Promise<void>): Promise<void> {
+    if (this.transactionDepth) {
+      this.preCommitCallbacks.push(fn);
+    } else {
+      // rely on thrown errors bubbling up from here to halt the transaction
+      await fn();
+    }
+  }
+
+  private async processPreCommit(): Promise<void> {
+    const callbacks = this.preCommitCallbacks;
+    this.preCommitCallbacks = [];
+    for (const cb of callbacks) {
+      // rely on thrown errors bubbling up from here to halt the transaction
+      await cb();
+    }
+  }
+
+  async postCommit(fn: () => Promise<void>): Promise<void> {
+    if (this.transactionDepth) {
+      this.postCommitCallbacks.push(fn);
+    } else {
+      await this.invokePostCommitCallback(fn);
+    }
+  }
+
+  private async processPostCommit(): Promise<void> {
+    const callbacks = this.postCommitCallbacks;
+    this.postCommitCallbacks = [];
+    for (const cb of callbacks) {
+      await this.invokePostCommitCallback(cb);
+    }
+  }
+
+  private async invokePostCommitCallback(fn: () => Promise<void>): Promise<void> {
+    try {
+      await fn();
+    } catch (err) {
+      if (err instanceof Error) {
+        getLogger().error('Error processing post-commit callback', err);
+      } else {
+        getLogger().error('Error processing post-commit callback', { err });
+      }
+    }
+  }
+
+  private pushCallbackFrame(): void {
+    this.callbackStack.push({
+      pre: this.preCommitCallbacks.length,
+      post: this.postCommitCallbacks.length,
+    });
+  }
+
+  private popCallbackFrame(): CallbackFrame {
+    const frame = this.callbackStack.pop();
+
+    if (!frame) {
+      throw new Error('No callback frame');
+    }
+
+    return frame;
+  }
+
+  private clearCallbackStack(): void {
+    this.callbackStack = [];
+  }
+
+  private truncateCommitCallbacks(): void {
+    const frame = this.popCallbackFrame();
+    if (frame.pre !== this.preCommitCallbacks.length) {
+      this.preCommitCallbacks = this.preCommitCallbacks.slice(0, frame.pre);
+    }
+    if (frame.post !== this.postCommitCallbacks.length) {
+      this.postCommitCallbacks = this.postCommitCallbacks.slice(0, frame.post);
+    }
+  }
+
+  [Symbol.dispose](removeConnection?: boolean): void {
+    this.assertNotClosed();
+    if (this.transactionDepth > 0) {
+      // Bad state, remove connection from pool
+      getLogger().error('Closing Repository with active transaction');
+      this.releaseConnection(new Error('Closing Repository with active transaction'));
+    } else {
+      // Good state, return healthy connection to pool
+      this.releaseConnection(removeConnection);
+    }
+    this.closed = true;
+  }
+
+  private assertNotClosed(): void {
+    if (this.closed) {
+      throw new Error('Already closed');
+    }
+  }
+}
